Target Words + WordNet Unofficial Code
this is a rough version of the code i originally used for getting the dataset we used for the experiment. i tried to recover it to the best of my ability, but there was a lot of override, repetition and variables that i lost. some parts are troubleshooted through chatgpt as a last resort.